Massive conversion from BigDecimal to long.
Now possible thanks to removing Qora v1 support.

Maximum asset quantities now unified to 10_000_000_000,
to 8 decimal places, removing prior 10 billion billion
indivisible maximum.

All values can now fit into a 64bit long.
(Except maybe when processing asset trades).

Added a general-use JAXB AmountTypeAdapter for converting
amounts to/from String/long.

Asset trading engine split into more methods for easier
readability.

Switched to using FIXED founder block reward distribution code,
ready for launch.

In HSQLDBDatabaseUpdates,
QortalAmount changed from DECIMAL(27, 0) to BIGINT
RewardSharePercent added to replace DECIMAL(5,2) with INT

Ripped out unused Transaction.isInvolved and Transaction.getAmount
in all subclasses.

Changed
  Transaction.getRecipientAccounts() : List<Account>
to
  Transaction.getRecipientAddresses() : List<String>
as only addresses are ever used.

Corrected returned values for above getRecipientAddresses() for
some transaction subclasses.

Added some account caching to some transactions to reduce repeated
loads during validation and then processing.

Transaction transformers:

Changed serialization of asset amounts from using 12 bytes to
now standard 8 byte long.

Updated transaction 'layouts' to reflect new sizes.

RewardShareTransactionTransformer still uses 8byte long to represent
reward share percent.

Updated some unit tests - more work needed!

+++ src/main/java/org/qortal/data/at/ATStateData.java
@@ -1,7 +1,5 @@
 package org.qortal.data.at;
 
-import java.math.BigDecimal;
-
 public class ATStateData {
 
 	// Properties
@@ -10,12 +8,12 @@ public class ATStateData {
 	private Long creation;
 	private byte[] stateData;
 	private byte[] stateHash;
-	private BigDecimal fees;
+	private Long fees;
 
 	// Constructors
 
 	/** Create new ATStateData */
-	public ATStateData(String ATAddress, Integer height, Long creation, byte[] stateData, byte[] stateHash, BigDecimal fees) {
+	public ATStateData(String ATAddress, Integer height, Long creation, byte[] stateData, byte[] stateHash, Long fees) {
 		this.ATAddress = ATAddress;
 		this.height = height;
 		this.creation = creation;
@@ -25,7 +23,7 @@ public class ATStateData {
 	}
 
 	/** For recreating per-block ATStateData from repository where not all info is needed */
-	public ATStateData(String ATAddress, int height, byte[] stateHash, BigDecimal fees) {
+	public ATStateData(String ATAddress, int height, byte[] stateHash, Long fees) {
 		this(ATAddress, height, null, null, stateHash, fees);
 	}
 
@@ -35,7 +33,7 @@ public class ATStateData {
 	}
 
 	/** For creating ATStateData from serialized bytes when we don't have all the info */
-	public ATStateData(String ATAddress, byte[] stateHash, BigDecimal fees) {
+	public ATStateData(String ATAddress, byte[] stateHash, Long fees) {
 		this(ATAddress, null, null, null, stateHash, fees);
 	}
 
@@ -66,7 +64,7 @@ public class ATStateData {
 		return this.stateHash;
 	}
 
-	public BigDecimal getFees() {
+	public Long getFees() {
 		return this.fees;
 	}
